Derivatives_Trade_system/requirements.txt
Django==2.2.6
django-crispy-forms==1.9.0
reportlab==3.5.34
scikit-learn==0.22.1
sklearn==0.0
tensorflow==2.0
Keras==2.3.1
djangorestframework==3.11.0
django-filter==2.2.0
django-simple-history==2.8.0
django-import-export==2.0.2
django-reporting==0.21

